test(patina): split require-toggle-inside-transition tests into sibling file
